02-Agents/09-Agent-State/20-autoread/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repository-analyzer"
version = "0.1.0"
description = "Repository structure analysis component for LangGraph agents"
readme = "README.md"
authors = [
    {name = "Kilo Code"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
requires-python = ">=3.8"
dependencies = [
    "langgraph>=0.1.0",
    "gitpython>=3.1.0",
    "pathlib2>=2.3.0",
    "pyyaml>=6.0.0",
    "toml>=0.10.0",
    "configparser>=5.0.0"
]

[project.optional-dependencies]
full = [
    "pygithub>=1.59.0",
    "chardet>=5.0.0"
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0"
]

[project.urls]
Homepage = "https://github.com/example/repository-analyzer"
Repository = "https://github.com/example/repository-analyzer"
Documentation = "https://github.com/example/repository-analyzer/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["repository_analyzer*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88